services/tts/qwen3-tts/requirements.txt
fastapi
uvicorn
python-dotenv
torchaudio
qwen-tts
scipy